requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==14.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
python-dotenv==1.0.1
alembic==1.14.0

